requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.23
Werkzeug==0.9.4
flake8==2.1.0
itsdangerous==0.24
mccabe==0.2.1
pep8==1.5.7
py==1.4.20
pyflakes==0.8.1
pytest==2.5.2
requests==2.3.0
wsgiref==0.1.2
gunicorn==19.3.0
flask-mysqldb==0.2.0
flask-httpauth==2.7.0
flask_sqlalchemy==2.1.0
alembic==0.8.4
Fabric==1.10.2
PyYAML==3.11.0
